src/main/java/com/VPI/VPI/Controllers/UsuarioController.java
package com.VPI.VPI.Controllers;


import com.VPI.VPI.Entities.UsuarioEntity;
import com.VPI.VPI.Services.UsuarioService;
import io.swagger.annotations.Api;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

//@PreAuthorize("authenticated")//seguridad basica del controlador
@RestController
@RequestMapping(value = "/usuarios")
@Api( tags = "Usuarios")
public class UsuarioController {

    @Autowired
    private UsuarioService usuarioService;

    @PreAuthorize("hasAuthority('Cliente')")
    @GetMapping("/ClientePage")
    public String ClientePage(){
        return ("<h1> Welcome Cliente</h1>");
    }

    @PreAuthorize("hasAuthority('Admin')")
    @GetMapping("/AdminPage")
    public String AdminPage(){
        return ("<h1> Welcome Admin</h1>");
    }

    @GetMapping("/Home")
    public String HomePage(){
        return ("<h1> Home</h1>");
    }


    @PreAuthorize("hasAuthority('UAdmin') OR hasAuthority('Restaurant')")
    @GetMapping("/GetAll")
    public List<UsuarioEntity> getAllUsuarios() {
        return usuarioService.getAllUsuarios();
    }

   
/*
    @RequestMapping(method = RequestMethod.GET, value = "/employees/{id}")
    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Long employeeId)
        throws ResourceNotFoundException {
        Employee employee = employeeRepository.findById(employeeId)
          .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
        return ResponseEntity.ok().body(employee);
    }
    


    @PutMapping("/employees/{id}")
    public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Long employeeId,
         @Valid @RequestBody Employee employeeDetails) throws ResourceNotFoundException {
        Employee employee = employeeRepository.findById(employeeId)
        .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));

        employee.setEmailId(employeeDetails.getEmailId());
        employee.setLastName(employeeDetails.getLastName());
        employee.setFirstName(employeeDetails.getFirstName());
        final Employee updatedEmployee = employeeRepository.save(employee);
        return ResponseEntity.ok(updatedEmployee);
    }

    @DeleteMapping("/employees/{id}")
    public Map<String, Boolean> deleteEmployee(@PathVariable(value = "id") Long employeeId)
         throws ResourceNotFoundException {
        Employee employee = employeeRepository.findById(employeeId)
       .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));

        employeeRepository.delete(employee);
        Map<String, Boolean> response = new HashMap<>();
        response.put("deleted", Boolean.TRUE);
        return response;
    }*/
    
}